requirements.txt
pandas==1.3.5
pyarrow==7.0.0
openai==1.3.0
pyyaml==6.0